removed vscode dependency from protocol-only
